@@ -1,6 +1,7 @@
-const defaultTheme = require("tailwindcss/defaultTheme");
+import defaultTheme from "tailwindcss/defaultTheme";
 
-module.exports = {
+/** @type {import('tailwindcss').Config} */
+export default {
   content: ["./src/**/*.{js,jsx,ts,tsx}"],
   theme: {
     extend: {
@@ -24,5 +25,6 @@ module.exports = {
       }
     }
   },
-  plugins: ["gatsby-plugin-postcss"]
-};
+  plugins: []
+}
+
